acr_autoscaler_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common
azure-core
azure-identity
azure-mgmt-containerregistry
azure-mgmt-core
certifi
cffi
cryptography
idna
isodate
msal
msal-extensions
pip
portalocker
pycparser
PyJWT
pywin32
requests
setuptools
six
typing_extensions
urllib3